Move base vmds to infrastructure
