+++ Core/Infrastructure/Stores/BaseVmdStore.cs
@@ -1,5 +1,5 @@
 using AppInfrastructure.Stores.DefaultStore;
-using Core.VMD.Base;
+using Core.Infrastructure.VMD;
 
 namespace Core.Infrastructure.Stores;
