+++ Core/Infrastructure/VMD/BaseMenuVmd.cs
@@ -1,6 +1,6 @@
 using ReactiveUI;
 
-namespace Core.VMD.Base;
+namespace Core.Infrastructure.VMD;
 
 public abstract class BaseMenuVmd<T> : BaseVmd,IMenuVmd<T>
 {
