+++ Core/VMD/DevPanelVmds/DevVmd.cs
@@ -1,5 +1,5 @@
 ﻿using Core.Infrastructure.Hosting;
-using Core.VMD.Base;
+using Core.Infrastructure.VMD;
 using Microsoft.Extensions.DependencyInjection;
